requirement.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
django-decouple==2.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
